core: pathfinding: move geometry stuff to a separate class
